+++ shenyu-client/shenyu-client-core/pom.xml
@@ -77,10 +77,6 @@
             <scope>provided</scope>
         </dependency>
         <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-autoconfigure</artifactId>
-        </dependency>
-        <dependency>
             <groupId>org.apache.shenyu</groupId>
             <artifactId>shenyu-client-api-docs-annotations</artifactId>
             <version>${project.version}</version>
