+++ shenyu-examples/shenyu-examples-http/src/main/java/org/apache/shenyu/examples/http/ShenyuTestHttpApplication.java
@@ -17,7 +17,7 @@
 
 package org.apache.shenyu.examples.http;
 
-import org.apache.shenyu.client.core.annotation.EnableClientRegister;
+import org.apache.shenyu.client.auto.config.EnableClientRegister;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
